# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Arena GameFi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/stake:
    post:
      operationId: stakeTokens
      tags: [wallet]
      summary: Stake ONE tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StakeRequest"
      responses:
        "200":
          description: Staking successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/unstake:
    post:
      operationId: unstakeTokens
      tags: [wallet]
      summary: Unstake ONE tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StakeRequest"
      responses:
        "200":
          description: Unstaking successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
...
components:
  schemas:
...
    Wallet:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        balance:
          type: number
        totalEarned:
          type: number
        totalSpent:
          type: number
        stakedAmount:
          type: number
        stakingRewards:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - balance
        - totalEarned
        - totalSpent
        - stakedAmount
        - stakingRewards
        - createdAt
        - updatedAt
...
    StakeRequest:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount
...